# internal/routes/transactions.go
package routes

import (
	"net/http"
	"strconv"

	"securewallet/internal/config"
	"securewallet/internal/middleware"
	"securewallet/internal/models"

	"github.com/gin-gonic/gin"
)

// SetupTransactionRoutes sets up transaction routes
func SetupTransactionRoutes(router *gin.RouterGroup) {
	transactions := router.Group("/transactions")
	{
		transactions.GET("", middleware.AuthMiddleware(), getTransactions)
		transactions.GET("/:id", middleware.AuthMiddleware(), getTransaction)
		transactions.POST("", middleware.AuthMiddleware(), createTransaction)
		transactions.PUT("/:id", middleware.AuthMiddleware(), updateTransaction)
		transactions.DELETE("/:id", middleware.AuthMiddleware(), deleteTransaction)
	}
}

// getTransactions gets all transactions for the current user
// @Summary List transactions
// @Description Get all transactions for the authenticated user's wallet
// @Tags transactions
// @Accept json
// @Produce json
// @Param limit query int false "Number of transactions to return" default(50)
// @Success 200 {array} models.Transaction
// @Failure 401 {object} gin.H
// @Failure 404 {object} gin.H
// @Security BearerAuth
// @Router /transactions [get]
func getTransactions(c *gin.Context) {
	user, exists := c.Get("user")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
		return
	}

	currentUser := user.(*models.User)
	db := config.GetDB()

	// Get user's wallet
	var userWallet models.Wallet
	if err := db.Where("user_id = ?", currentUser.ID).First(&userWallet).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Wallet not found"})
		return
	}

	// Get limit parameter
	limitStr := c.Query("limit")
	limit := 50 // default limit
	if limitStr != "" {
		if parsedLimit, err := strconv.Atoi(limitStr); err == nil && parsedLimit > 0 {
			limit = parsedLimit
		}
	}

	// Get transactions for the user's wallet
	var transactions []models.Transaction
	if err := db.Where("wallet_id = ?", userWallet.ID).
		Order("created_at DESC").
		Limit(limit).
		Find(&transactions).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to fetch transactions"})
		return
	}

	c.JSON(http.StatusOK, transactions)
}

// getTransaction gets a specific transaction
// @Summary Get transaction by ID
// @Description Get a specific transaction by ID
// @Tags transactions
// @Accept json
// @Produce json
// @Param id path string true "Transaction ID"
// @Success 200 {object} gin.H
// @Security BearerAuth
// @Router /transactions/{id} [get]
func getTransaction(c *gin.Context) {
	id := c.Param("id")
	c.JSON(http.StatusOK, gin.H{"message": "Get transaction", "id": id})
}

// createTransaction creates a new transaction
// @Summary Create transaction
// @Description Create a new transaction record
// @Tags transactions
// @Accept json
// @Produce json
// @Success 201 {object} gin.H
// @Security BearerAuth
// @Router /transactions [post]
func createTransaction(c *gin.Context) {
	c.JSON(http.StatusCreated, gin.H{"message": "Create transaction"})
}

// updateTransaction updates a transaction
// @Summary Update transaction
// @Description Update a transaction record
// @Tags transactions
// @Accept json
// @Produce json
// @Param id path string true "Transaction ID"
// @Success 200 {object} gin.H
// @Security BearerAuth
// @Router /transactions/{id} [put]
func updateTransaction(c *gin.Context) {
	id := c.Param("id")
	c.JSON(http.StatusOK, gin.H{"message": "Update transaction", "id": id})
}

// deleteTransaction deletes a transaction
// @Summary Delete transaction
// @Description Delete a transaction record
// @Tags transactions
// @Accept json
// @Produce json
// @Param id path string true "Transaction ID"
// @Success 200 {object} gin.H
// @Security BearerAuth
// @Router /transactions/{id} [delete]
func deleteTransaction(c *gin.Context) {
	id := c.Param("id")
	c.JSON(http.StatusOK, gin.H{"message": "Delete transaction", "id": id})
}